requirements.txt
pyyaml>=6.0.1
docker>=7.0.0
loguru>=0.7.0
psutil>=7.0.0